requirements.txt
# Core dependencies
requests>=2.28.0
openai>=1.0.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Environment
python-dotenv>=1.0.0

# OAuth
authlib>=1.3.0
httpx>=0.27.0

# Encryption
cryptography>=41.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
